ref: Refactor component and move to the root page + adding refreshing after uploading
