+++ app/page.tsx
@@ -1,15 +1,39 @@
-"use client";
-import { CldImage, CldUploadButton } from "next-cloudinary";
+import { v2 as cloudinary } from "cloudinary";
+import CCldImage from "./Components/CCldImage";
+import UploadButton from "./Components/UploadButton";
 
-export default function Home() {
+type imageData = {
+  public_id: string;
+};
+
+export default async function GalleryPage() {
+  const { resources } = await cloudinary.search
+    .expression("resource_type:image")
+    .sort_by("created_at", "desc")
+    .max_results(30)
+    .execute();
+
   return (
     <main>
-      <CldImage
-        width={900}
-        height={900}
-        src="https://res.cloudinary.com/ddomuvkaw/image/upload/v1704475945/cld-sample-5.jpg"
-        alt="Cloudinary Sample Image"
-      />
+      <div className="flex justify-between items-center mb-6">
+        <h1 className="text-3xl">Gallery page</h1>
+        <UploadButton />
+        <button className="bg-lime-600 text-white p-2 rounded">
+          <a href="/link">Create link</a>
+        </button>
+      </div>
+      <div className="grid grid-cols-4 gap-4">
+        {resources &&
+          resources.map((image: imageData) => (
+            <CCldImage
+              src={image.public_id}
+              alt="image"
+              height={100}
+              width={100}
+              key={image.public_id}
+            />
+          ))}
+      </div>
     </main>
   );
 }
